+++ BBus/BBus/Global/BusPosByRtidDTO.swift
@@ -1,0 +1,30 @@
+//
+//  BusPosByRtid.swift
+//  BBus
+//
+//  Created by 최수정 on 2021/11/10.
+//
+
+import Foundation
+
+//struct BusPosByRtidDTO: BBusXMLDTO {
+//    let busType: Int
+//    let congestion: Int
+//    let plainNumber: String
+//    let sectionOrder: Int
+//    
+//    init?(dict: [String : [Any]]) {
+//        guard let busTypeString = dict["busType"],
+//              let busType = Int(busTypeString),
+//              let congestionString = dict["congetion"],
+//              let congestion = Int(congestionString),
+//              let plainNumber = dict["plainNo"],
+//              let sectOrd = dict["sectOrd"],
+//              let sectionOrder = Int(sectOrd) else { return nil }
+//        
+//        self.busType = busType
+//        self.congestion = congestion
+//        self.plainNumber = plainNumber
+//        self.sectionOrder = sectionOrder
+//    }
+//}
