+++ BBus/BBus/Global/StationByUidItemDTO.swift
@@ -1,0 +1,26 @@
+//
+//  StationByUidItem.swift
+//  BBus
+//
+//  Created by 최수정 on 2021/11/10.
+//
+
+import Foundation
+
+//struct StationByUidItemDTO: BBusXMLDTO {
+//    let sectionSpeed: Int
+//    let sequence: Int
+//    let stationName: String
+//
+//    init?(dict: [String : [Any]]) {
+//        guard let sectSpd = dict["sectSpd"],
+//              let sectionSpeed = Int(sectSpd),
+//              let seq = dict["seq"],
+//              let sequence = Int(seq),
+//              let stationName = dict["stationNm"] else { return nil }
+//
+//        self.sectionSpeed = sectionSpeed
+//        self.sequence = sequence
+//        self.stationName = stationName
+//    }
+//}
